format sql

+++ models/examples/seed_model.sql
@@ -4,8 +4,8 @@ MODEL (
     path '../../seeds/seed_data.csv'
   ),
   columns (
-    id INT,
-    item_id INT,
+    id INT64,
+    item_id INT64,
     event_date DATE
   ),
   grain (id, event_date)
